Simplifying script
