+++ constants.py
@@ -1,5 +1,4 @@
 VERSION = 0.1
-AUTHOR = ""
-APP_NAME = "Chatter"
+APP_NAME = "Cleaner"
 PLATFORM = "linux"
 USER_AGENT = f"{PLATFORM}:{APP_NAME}:v{VERSION} (by u/{AUTHOR})"
